requirements.txt
flet
langchain
openai
pandas
python-dotenv
pinecone-client
tiktoken
pypdf